requirements.txt
args==0.1.0
certifi==2020.12.5
chardet==4.0.0
click==6.7
clint==0.5.1
cycler==0.10.0
e2cnn==0.1.5
idna==2.10
kiwisolver==1.3.1
marshmallow==2.21.0
matplotlib==3.3.4
numpy==1.20.1
pathlib2==2.3.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
raven==6.10.0
requests==2.25.1
requests-toolbelt==0.9.1
scipy==1.6.1
six==1.15.0
tabulate==0.8.9
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.3
